프래그먼트 생성 후 상단 내비게이션바와 연결
