+++ app/src/main/java/com/ajou/ourvillage/Main/MainFragment.java
@@ -1,0 +1,25 @@
+package com.ajou.ourvillage.Main;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.ajou.ourvillage.R;
+
+public class MainFragment extends Fragment {
+
+    public MainFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_main, container, false);
+    }
+}
