+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    android:id="@+id/main_root"
+    android:id="@+id/root"
     tools:context=".MainActivity">
 
     <com.google.android.material.appbar.AppBarLayout
